ch3-bank-transaction-analyzer-extends/src/main/java/com/icloud/parser/BankStatementCSVParser.java
package com.icloud.parser;

import com.icloud.model.BankTransaction;

import java.time.LocalDate;
import java.time.format.DateTimeFormatter;

public class BankStatementCSVParser implements BankStatementParser {
    private static final DateTimeFormatter DATE_PATTERN
            = DateTimeFormatter.ofPattern("dd-MM-yyyy");

    @Override
    public BankTransaction parseFrom(final String line) {
        String[] columns = line.split(",");

        final LocalDate date = LocalDate.parse(columns[0], DATE_PATTERN);
        final double amount = Double.parseDouble(columns[1]);
        final String description = columns[2];

        return new BankTransaction(date, amount, description);
    }

}
